+++ source/Core/HELM/Graph.h
@@ -20,7 +20,7 @@ public:
 	std::vector<int> calculateReverseCuthillMcKee(int startNode) const;
 	std::vector<std::vector<int>> createLayeringFrom(int startNode) const;
 	int findPseudoPeriphereNode() const;
-	int findPseudoPeriphereNode(int &eccentricity) const;
+	int findPseudoPeriphereNode(size_t &eccentricity) const;
 	int calculateSecondLevelDegree(int index) const;
 
 private:
